Типы в библиотеках

+++ module2/6.1/main.c
@@ -157,7 +157,7 @@ void print_address(const address *a)
 
 void print_contact(const contact *c)
 {
-    printf("Контакт #%llu:\n", c->id);
+    printf("Контакт #%lu:\n", c->id);
 
     print_names(&c->names);
     print_address(&c->address);
@@ -326,8 +326,8 @@ void menu2(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     if (delete_contact_from_phonebook(pb, id, 1) == 0)
     {
@@ -348,8 +348,8 @@ void menu3(phonebook *pb)
     }
 
     printf("Введите id контакта для редактирования: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     printf("Форматные строки для редактирования:\n");
     printf("1.1 - фамилия, 1.2 - имя, 1.3 - отчество\n");
@@ -421,8 +421,8 @@ void menu5(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -459,8 +459,8 @@ void menu6(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -496,8 +496,8 @@ void menu7(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -533,8 +533,8 @@ void menu8(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -573,8 +573,8 @@ void menu9(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -613,8 +613,8 @@ void menu10(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -653,8 +653,8 @@ void menu11(phonebook *pb)
     }
 
     printf("Введите id контакта для поиска: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -681,7 +681,7 @@ void menu12(phonebook *pb)
     char last_name[LAST_NAME_LENGTH];
     scanf("%s", last_name);
 
-    unsigned int count;
+    uint64_t count;
     contact **cs = find_by_last_name(pb, last_name, &count);
 
     if (cs == NULL)
@@ -691,7 +691,7 @@ void menu12(phonebook *pb)
     else
     {
         printf("Найдены контакты:\n");
-        for (unsigned int i = 0; i < count; i++)
+        for (uint64_t i = 0; i < count; i++)
         {
             print_contact(cs[i]);
         }
